No issue reporting for unsupported add-ons, busy dialog box fix
Disallow reporting for removed ten and plus7 add-ons.

The old method of showing a busy dialog box seems to not work anymore. Reports sometimes get double or triple reported as you can start a new one while the old is running in the background. Now uses the progress dialog box - there is a 'busy' dialog box type but that was only introduced in Krypton.