Issues not allowed to be submitted unless url is from current version
Users need not be on the latest version of an addon (allows for dev versions) but can only submit an issue report if the url they're accessing was created with the same version of the addon they're using. This stops the constant flow of reports caused by using an old 'favourite' listitem.